pyproject.toml
[tool.poetry]
name = "vue-fastapi-postgresql"
version = "0.0.1"
description = "Full Stack FastAPI and PostgreSQL - Base Project Generator"
authors = ["Vlad Emelianov <vlad.emelianov.nz@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/youtype/vue-fastapi-postgresql"
repository = "https://github.com/youtype/vue-fastapi-postgresql"
documentation = "https://github.com/youtype/vue-fastapi-postgresql"
keywords = ["vue", "fastapi", "postgresql", "cookiecutter"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
    "Typing :: Typed",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [
]
include = [
]

[tool.poetry.urls]
"Documentation" = "https://github.com/youtype/vue-fastapi-postgresql"
"Source" = "https://github.com/youtype/vue-fastapi-postgresql"
"Bug Tracker" = "https://github.com/youtype/vue-fastapi-postgresql/issues"

[tool.poetry.dependencies]
python = "^3.7"
pip = "*"
uvicorn = "^0.18.3"
fastapi = "^0.82.0"
python-multipart = "^0.0.5"
email-validator = "^1.2.1"
requests = "^2.28.1"
celery = "^5.2.7"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^8.0.1"
pydantic = "^1.10.2"
emails = "^0.6"
raven = "^6.10.0"
gunicorn = "^20.1.0"
jinja2 = "^3.1.2"
psycopg2-binary = "^2.9.3"
alembic = "^1.8.1"
sqlalchemy = "^1.4.40"
pytest = "^7.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sentry-sdk = "^1.12.1"
asyncpg = "^0.27.0"
pytest-asyncio = "^0.20.3"
httpx = "^0.23.1"

[tool.poetry.dev-dependencies]
handsdown = "*"
mypy = "*"
black = "*"
isort = "*"
flake8 = "*"
pytest = "*"
pyyaml = "*"
types-python-jose = "*"
types-passlib = "*"
celery-types = "*"
sqlalchemy2-stubs = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
include = '(mypy_boto3_builder|types_s3transfer|tests)/.*\.pyi?$'
target-version = [
    'py37',
    'py38',
    'py39',
    'py310',
    'py311',
]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = [
    "app",
    "tests",
]
src_paths = []

[tool.coverage.report]
exclude_lines = [
    "if self.debug:",
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@overload",
]
ignore_errors = true
omit = [
    "tests/*",
]

[tool.pyright]
exclude = [
    "**/__pycache__",
    "tests",
    "**/build",
    "typestubs",
]
# typeCheckingMode = "strict"
reportMissingImports = "error"
reportMissingTypeStubs = "error"
reportMissingTypeArgument = "error"
reportIncompatibleMethodOverride = "error"
reportIncompatibleVariableOverride = "error"
reportUnknownParameterType = "error"
pythonVersion = "3.7"
